+++ test/platform/swift/unit_integration/mocks/MockLoggerBridging.swift
@@ -144,5 +144,5 @@ extension MockLoggerBridging: LoggerBridging {
 
     public func setFeatureFlagExposure(withName flag: String, variant: String) {}
 
-    public func registerOpaqueEntityID(_: String) {}
+    public func setEntityID(_: String) {}
 }